% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceteris_paribus.R
\name{ceteris_paribus}
\alias{ceteris_paribus}
\title{Ceteris Paribus Plot}
\usage{
ceteris_paribus(explainer, observation, selected_variables = NULL)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'survxai::explain' function}

\item{observation}{a new observarvation for which predictions need to be explained}

\item{selected_variables}{if specified, then only these variables will be explained}
}
\value{
An object of the class surv_ceteris_paribus_explainer.
It's a data frame with calculated average responses.
}
\description{
Ceteris Paribus Plot
}
