% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcTest.R
\docType{data}
\name{pbcTest}
\alias{pbcTest}
\title{pbcTest}
\source{
randomForestSRC
}
\description{
PBC test set
Data set based on \code{pbc} from \code{randomForestSRC} package. 
The data consists of 138 randomly chosen observations The \code{pbcTest} contains only complete cases for each observation.
It contains 5 variables: `status`, `sex`, `bili`, `stage`, and `years`.
}
\references{
Flemming T.R and Harrington D.P., (1991) Counting Processes and Survival Analysis. New York: Wiley.
}
