% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcTrain.R
\docType{data}
\name{pbcTrain}
\alias{pbcTrain}
\title{pbcTrain}
\source{
randomForestSRC
}
\description{
PBC train set
Data set based on \code{pbc} from \code{randomForestSRC} package. 
The data consists of 138 randomly chosen observations The \code{pbcTrain} contains only complete cases for each observation.
It contains 5 variables: `status`, `sex`, `bili`, `stage`, and `years`.
}
\examples{
data("pbcTrain", package = "survxai")
head(pbcTrain)
}
\references{
Flemming T.R and Harrington D.P., (1991) Counting Processes and Survival Analysis. New York: Wiley.
}
