% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv_breakdown.R
\name{plot.surv_breakdown}
\alias{plot.surv_breakdown}
\title{Plot for surv_breakdown object}
\usage{
\method{plot}{surv_breakdown}(x, table = FALSE, ...)
}
\arguments{
\item{x}{object of class "surv_breakdown"}

\item{table}{if we want to plot the table of contributions}

\item{...}{other arguments}
}
\description{
Function plot for surv_breakdown object visualise estimated survival curve of mean probabilities in chosen time points.
}
