% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ceteris_paribus.R
\name{plot.surv_ceteris_paribus_explainer}
\alias{plot.surv_ceteris_paribus_explainer}
\title{Plot for ceteris_paribus object}
\usage{
\method{plot}{surv_ceteris_paribus_explainer}(x, ...)
}
\arguments{
\item{x}{object of class "surv_ceteris_paribus_explainer"}

\item{...}{other arguments}
}
\description{
Function plot for ceteris_paribus object visualise estimated survival curve of mean probabilities in chosen time points.
}
