% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explainer.R
\name{plot.surv_explainer}
\alias{plot.surv_explainer}
\title{Plot for surv_explainer object}
\usage{
\method{plot}{surv_explainer}(x, ...)
}
\arguments{
\item{x}{object of class "surv_explainer"}

\item{...}{other arguments for function \code{\link[survminer]{ggsurvplot}}}

\item{times}{numeric vector of time points}
}
\description{
Function plot for surv_explainer object visualise estimated survival curve of mean probabilities in chosen time points.
}
