% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_performance.R
\name{plot.surv_model_performance_explainer}
\alias{plot.surv_model_performance_explainer}
\title{Plot for surv_model_performance object}
\usage{
\method{plot}{surv_model_performance_explainer}(x, ...)
}
\arguments{
\item{x}{object of class "surv_model_performance"}

\item{...}{other arguments}
}
\description{
Function plot for surv_model_perfomance object.
}
\examples{
\dontrun{
   library(survxai)
   library(randomForestSRC)
   library(rms)
   data(pbc, package = "randomForestSRC")
   pbc <- pbc[complete.cases(pbc),]
   cph_model <- cph(Surv(days/365, status)~., data=pbc, surv=TRUE, x = TRUE, y=TRUE)
   surve_cph <- explain(model = cph_model, data = pbc[,-c(1,2)], y = Surv(pbc$days/365, pbc$status))
   mp_cph <- model_performance(surve_cph, data = pbc, reference_formula = Surv(days/365, status)~1)
   plot(mp_cph)
}

}
