% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction_breakdown.R
\name{plot.surv_prediction_breakdown_explainer}
\alias{plot.surv_prediction_breakdown_explainer}
\title{Plot for surv_breakdown object}
\usage{
\method{plot}{surv_prediction_breakdown_explainer}(x, ...)
}
\arguments{
\item{x}{object of class "surv_prediction_breakdown_explainer"}

\item{...}{other arguments}
}
\description{
Function plot for surv_breakdown object visualise estimated survival curve of mean probabilities in chosen time points.
}
\examples{
\dontrun{ 
library(survxai)
library(rms) 
library(randomForestSRC)
data(pbc, package = "randomForestSRC")
pbc <- pbc[complete.cases(pbc),]
predict_times <- function(model, data, times){ 
                  prob <- rms::survest(model, data, times = times)$surv
                  return(prob)
                  }
cph_model <- cph(Surv(days/365, status)~., data=pbc, surv=TRUE, x = TRUE, y=TRUE)
surve_cph <- explain(model = cph_model, data = pbc[,-c(1,2)], y = Surv(pbc$days/365, pbc$status), 
             predict_function = predict_times)
broken_prediction <- prediction_breakdown(surve_cph, pbc[1,-c(1,2)])
plot(broken_prediction)
}
}
