% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv_variable_response.R
\name{plot.surv_variable_response}
\alias{plot.surv_variable_response}
\title{Plot for surv_variable_serponse object}
\usage{
\method{plot}{surv_variable_response}(x, ...)
}
\arguments{
\item{x}{object of class "surv_variable_response"}

\item{...}{other arguments}
}
\description{
Function plot for surv_variable_response object shows the expected output condition on a selected variable.
}
