% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_response.R
\name{plot.surv_variable_response_explainer}
\alias{plot.surv_variable_response_explainer}
\title{Plot for surv_variable_response object}
\usage{
\method{plot}{surv_variable_response_explainer}(x, ..., split = "model")
}
\arguments{
\item{x}{object of class "surv_variable_response"}

\item{...}{other arguments}

\item{split}{a character, either "model" or "variable". Sets the variable for faceting.}
}
\description{
Function plot for surv_variable_response object shows the expected output condition on a selected variable.
}
\examples{
\dontrun{
library(survxai)
library(rms)
library(randomForestSRC)
data(pbc, package = "randomForestSRC")
pbc <- pbc[complete.cases(pbc),]
predict_times <- function(model, data, times){
                  prob <- rms::survest(model, data, times = times)$surv
                  return(prob)
                  }
cph_model <- cph(Surv(days/365, status)~., data=pbc, surv=TRUE, x = TRUE, y=TRUE)
surve_cph <- explain(model = cph_model, data = pbc[,-c(1,2)], y = Surv(pbc$days/365, pbc$status),
             predict_function = predict_times)
svr_cph <- variable_response(surve_cph, "sex")
plot(svr_cph)
}

}
