% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_breakdown.R
\name{prediction_breakdown}
\alias{prediction_breakdown}
\title{BreakDown for survival models}
\usage{
prediction_breakdown(explainer, observation, ...)
}
\arguments{
\item{explainer}{an object of the class 'surv_explainer'}

\item{observation}{a new observation to explain}

\item{...}{other parameters}
}
\value{
An object of class surv_prediction_breakdown_explainer
}
\description{
Function \code{surv_breakdown} is an extension of a broken function from breakDown package.
}
