% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ceteris_paribus.R
\name{print.surv_ceteris_paribus_explainer}
\alias{print.surv_ceteris_paribus_explainer}
\title{Ceteris Paribus Print}
\usage{
\method{print}{surv_ceteris_paribus_explainer}(x, ...)
}
\arguments{
\item{x}{the model of 'surv_ceteris_paribus_explainer' class}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data frame
}
\description{
Ceteris Paribus Print
}
