% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_explainer.R
\name{print.surv_explainer}
\alias{print.surv_explainer}
\title{Print Survival Explainer Summary}
\usage{
\method{print}{surv_explainer}(x, ...)
}
\arguments{
\item{x}{a model survival expaliner created with the `explain()` function}

\item{...}{other parameters}
}
\description{
Print Survival Explainer Summary
}
