% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_performance.R
\name{print.surv_model_performance_explainer}
\alias{print.surv_model_performance_explainer}
\title{Print Survival Model Performance}
\usage{
\method{print}{surv_model_performance_explainer}(x, times = NULL, ...)
}
\arguments{
\item{x}{a model to be explained, object of the class 'model_performance_explainer'}

\item{times}{a vector of integer times on which we want to check the value of prediction error}

\item{...}{other parameters}
}
\description{
Print Survival Model Performance
}
