% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_prediction_breakdown.R
\name{print.surv_prediction_breakdown_explainer}
\alias{print.surv_prediction_breakdown_explainer}
\title{Prediction Breakdown Print}
\usage{
\method{print}{surv_prediction_breakdown_explainer}(x, ..., digits = 3,
  rounding_function = round)
}
\arguments{
\item{x}{the model model of 'surv_prediction_breakdown_explainer' class}

\item{...}{other parameters}

\item{digits}{number of decimal places (round) or significant digits (signif) to be used
See the \code{rounding_function} argument}

\item{rounding_function}{function that is to used for rounding numbers.
It may be \code{signif()} which keeps a specified number of significant digits.
Or the default \code{round()} to have the same precision for all components}
}
\description{
Prediction Breakdown Print
}
