% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_breakdown.R
\name{surv_breakdown}
\alias{surv_breakdown}
\title{BreakDown for survival models}
\usage{
surv_breakdown(explainer, new_observation, times = 1)
}
\arguments{
\item{explainer}{an object of the class 'surv_explainer'}

\item{new_observation}{a new observation to explain}

\item{times}{numeric vector with event times.}
}
\description{
Function \code{surv_breakdown} is an extension of a broken function from breakDown package.
}
