% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_explain.R
\name{surv_explain.default}
\alias{surv_explain.default}
\alias{surv_explain}
\title{Create Survival Model Explainer}
\usage{
surv_explain.default(model, data = NULL, time = NULL, status = NULL,
  predict_function = yhat, ..., label = tail(class(model), 1))

surv_explain(model, data = NULL, time = NULL, status = NULL,
  predict_function = yhat, ..., label = tail(class(model), 1))
}
\arguments{
\item{model}{object - a survival model to be explained}

\item{data}{data.frame or matrix - data that will be used by survival explainers. If not provided then will be extracted from the model}

\item{time}{numeric vector with event times.}

\item{status}{numeric vector with event status.}

\item{predict_function}{function that takes two arguments: model, new data, and times and returns numeric vector or matrix with predictions.}

\item{...}{other parameters}

\item{label}{character - the name of the survival model. By default it's extracted from the 'class' attribute of the model.}
}
\value{
An object of the class 'surv_explainer'.

It's a list with following fields:

\itemize{
\item \code{model} the explained model
\item \code{data} the dataset
\item \code{time} event times
\item \code{status} event statuses
\item \code{predict_function} function that may be used for model predictions, shall return a single numerical value for each observation.
\item \code{class} class/classes of a model
\item \code{label} label, by default it's the last value from the \code{class} vector, but may be set to any character.
}
}
\description{
Survival models may have very different structures.
This function creates a unified representation of a survival model, which can be further processed by various survival explainers.

Please NOTE, that the \code{model} is actually the only required argument.
But some survival explainers may require that others will be provided too.
}
