% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_variable_response.R
\name{surv_variable_response}
\alias{surv_variable_response}
\title{Variable response for survival models}
\usage{
surv_variable_response(explainer, variable)
}
\arguments{
\item{explainer}{an object of the class 'surv_explainer'.}

\item{variable}{character with variable name.}
}
\description{
Function \code{surv_variable response} calculates the expected output condition on a selected variable.
}
