% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mi2.R
\name{theme_mi2}
\alias{theme_mi2}
\title{MI^2 plot theme}
\usage{
theme_mi2()
}
\value{
theme object that can be added to ggplot2 plots
}
\description{
ggplot theme for charts generated with MI^2 Data Lab packages.
}
