% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_response.R
\name{variable_response}
\alias{variable_response}
\title{Variable response for survival models}
\usage{
variable_response(explainer, variable, type = "pdp", trans = explainer$link,
  ...)
}
\arguments{
\item{explainer}{an object of the class 'surv_explainer'.}

\item{variable}{character with variable name.}

\item{type}{character - type of the response to be calculated.
Currently following options are implemented: 'pdp' for Partial Dependency.}

\item{trans}{function - a transformation/link function that shall be applied to raw model predictions. This will be inherited from the explainer.}

\item{...}{other parameters}
}
\description{
Function \code{variable response} calculates the expected output condition on a selected variable.
}
