% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Model performance for survival models}
\usage{
model_performance(explainer, type = "BS", data = NULL,
  reference_formula = NULL, ...)
}
\arguments{
\item{explainer}{an object of the class 'surv_explainer'.}

\item{type}{character - type of the response to be calculated.
Currently following options are implemented: 'BS' for Expected Brier Score.}

\item{data}{data with time and status.}

\item{reference_formula}{A survival formula as obtained either with prodlim::Hist or survival::Surv. The left hand side is used to find the status response variable in data. For right censored data, the right hand side of the formula is used to specify conditional censoring models. As in \code{\link[pec]{pec}}.}

\item{...}{other parameters}
}
\description{
Function \code{model_performance} calculates the residuals and model performance measures.
}
\examples{
\dontrun{
   library(survxai)
   library(rms)
   library(randomForestSRC)
   data(pbc, package = "randomForestSRC")
   pbc <- pbc[complete.cases(pbc),]
   cph_model <- cph(Surv(days/365, status)~., data=pbc, surv=TRUE, x = TRUE, y=TRUE)
   surve_cph <- explain(model = cph_model, data = pbc[,-c(1,2)], y = Surv(pbc$days/365, pbc$status))
   mp_cph <- model_performance(surve_cph, data = pbc, reference_formula = Surv(days/365, status)~1)
}

}
\references{
Ulla B. Mogensen, Hemant Ishwaran, Thomas A. Gerds (2012). Evaluating Random Forests for Survival Analysis Using Prediction Error Curves. Journal of Statistical Software, 50(11), 1-23. URL http://www.jstatsoft.org/v50/i11/.
}
