% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explainer.R
\name{plot.surv_explainer}
\alias{plot.surv_explainer}
\title{Plot for surv_explainer object}
\usage{
\method{plot}{surv_explainer}(x, ...)
}
\arguments{
\item{x}{object of class "surv_explainer"}

\item{...}{other arguments for function \code{\link[survminer]{ggsurvplot}}}
}
\description{
Function plot for surv_explainer object visualise estimated survival curve of mean probabilities in chosen time points.
}
\examples{
\dontrun{
library(survxai)
library(rms) 
library(randomForestSRC)
data(pbc, package = "randomForestSRC")
pbc <- pbc[complete.cases(pbc),]
predict_times <- function(model, data, times){ 
                  prob <- rms::survest(model, data, times = times)$surv
                  return(prob)
                  }
cph_model <- cph(Surv(days/365, status)~., data=pbc, surv=TRUE, x = TRUE, y=TRUE)
surve_cph <- explain(model = cph_model, data = pbc[,-c(1,2)], y = Surv(pbc$days/365, pbc$status), 
             predict_function = predict_times)
plot(surve_cph)
}
}
