% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explainer.R
\name{plot.surv_explainer}
\alias{plot.surv_explainer}
\title{Plot for surv_explainer object}
\usage{
\method{plot}{surv_explainer}(x, times = 1, ...)
}
\arguments{
\item{x}{object of class "surv_explainer"}

\item{times}{numeric vector of time points}

\item{...}{other arguments}
}
\description{
Function plot for surv_explainer object visualise estimated survival curve of mean probabilities in chosen time points.
}
