% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction_breakdown.R
\name{plot.surv_prediction_breakdown_explainer}
\alias{plot.surv_prediction_breakdown_explainer}
\title{Plot for surv_breakdown object}
\usage{
\method{plot}{surv_prediction_breakdown_explainer}(x, ...)
}
\arguments{
\item{x}{object of class "surv_prediction_breakdown_explainer"}

\item{...}{other arguments}
}
\description{
Function plot for surv_breakdown object visualise estimated survival curve of mean probabilities in chosen time points.
}
