% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_response.R
\name{plot.surv_variable_response_explainer}
\alias{plot.surv_variable_response_explainer}
\title{Plot for surv_variable_serponse object}
\usage{
\method{plot}{surv_variable_response_explainer}(x, ..., split = "model")
}
\arguments{
\item{x}{object of class "surv_variable_response"}

\item{...}{other arguments}

\item{split}{a character, either "model" or "level". Sets the variable for faceting.}
}
\description{
Function plot for surv_variable_response object shows the expected output condition on a selected variable.
}
