% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_variable_response.R
\name{print.surv_variable_response_explainer}
\alias{print.surv_variable_response_explainer}
\title{Variable Response Print}
\usage{
\method{print}{surv_variable_response_explainer}(x, ...)
}
\arguments{
\item{x}{the model of 'surv_variable_response_explainer' class}

\item{...}{other parameters}
}
\value{
a data frame
}
\description{
Variable Response Print
}
